airflow/requirements.txt
apache-airflow-providers-google>=10.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
requests>=2.31.0
